+++ Smarts.Core/CollectionExtensions.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+
+namespace Smarts.Core
+{
+    public static class CollectionExtensions
+    {
+        public static void AddRange<T>(this ICollection<T> collection, IEnumerable<T> values)
+        {
+            foreach (var value in values)
+            {
+                collection.Add(value);
+            }
+        }
+    }
+}