+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.2.0
 - 🆕 `AsForeground` and `AsBackground` extension methods to change console
   colours with the `IDisposable` pattern resetting the previous colour.
+- 🆕 `AddRange` extension method bringing the functionality of `List<T>`'s
+  method to all `ICollection<T>`.
 
 # 0.1.0
 - 🆕 `Disposable` class for managing `IDisposable` instances.